+++ src/index.ts
@@ -5,6 +5,17 @@ import WinningHand from './hand/WinningHand';
 import Hand from './hand/Hand';
 import ExplorerOfWinningPermutations from './breadthFirstSearch/ExplorerOfWinningPermutations';
 import FaanCalculator from './calculateFaan/FaanCalculator';
+import FaanCalculationConfig from './calculateFaan/FaanCalculationConfig';
 import HandTypeFinder from './hand/handType/HandTypeFinder';
 
-export { WinningHand, Hand, FaanCalculator, Meld, meldFormer, Tile, ExplorerOfWinningPermutations, HandTypeFinder };
+export {
+  WinningHand,
+  Hand,
+  FaanCalculator,
+  FaanCalculationConfig,
+  Meld,
+  meldFormer,
+  Tile,
+  ExplorerOfWinningPermutations,
+  HandTypeFinder,
+};